http-solution/requirements.txt
# 使用原生 MCP SDK，避免版本兼容性问题
mcp==1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
uvicorn>=0.24.0
pydantic>=2.5.0
PyYAML>=6.0.1
Jinja2>=3.1.2
